trying to fix drag gesture

+++ Sources/SwiftCharts/LineChart/LineChart+View.swift
@@ -72,7 +72,7 @@ private extension SwiftCharts.LineChart.ChartView {
                 print(geometry.size.height - points[index].y)
                 print(points[index].y)
                 
-                indicatorOffset = .init(width: points[index].x - 40, height: -(geometry.size.height - points[index].y + config.extraHeadSpace))//points[index].y )//- geometry.size.height)
+                indicatorOffset = .init(width: points[index].x - 40, height: -(geometry.size.height - points[index].y + CGFloat(config.extraHeadSpace)))//points[index].y )//- geometry.size.height)
             }).onEnded({ value in
                 withAnimation {
                     showIndicator = false
